Module crrl::ed25519

Edwards25519 curve implementation.

This module implements generic group operations on the twisted Edwards curve of equation -x^2 + y^2 = 1 + d*x^2*y^2, over the finite field GF(2^255 - 19), for the constant d = -121665/121666. This curve is described in RFC 7748. The signature algorithm Ed25519, which operates on that curve, is described in RFC 8032.

The curve has order 8*L for a given prime integer L (which is slightly greater than 2^252). A conventional base point is defined, that generates the subgroup of order L. Points used in public keys and signatures are normally part of that subgroup, though this is not verified (the function Point::is_in_subgroup() can be used to validate that a given point is in that subgroup).

A curve point is represented by the Point structure. Point instances can be used in additions and subtractions with the usual + and - operators; all combinations of raw values and references are accepted, as well as compound assignment operators += and -=. Specialized functions are available, in particular for point doubling (Point::double()) and for sequences of successive doublings (Point::xdouble()), the latter using some extra optimizations. Multiplication by an integer (u64 type) or a scalar (Scalar structure) is also accepted, using the * and *= operators. Scalars are integers modulo L. The Scalar structure represents such a value; it implements all usual arithmetic operators (+, -, * and /, as well as +=, -=, *= and /=).

Scalars can be encoded over 32 bytes (using unsigned little-endian convention) and decoded back. Encoding is always canonical, and decoding always verifies that the value is indeed in the canonical range.

Points can be encoded over 32 bytes, and decoded back. As with scalars, encoding is always canonical, and verified upon decoding.

The PrivateKey structure represents a private key for the Ed25519 signature algorithm. It is instantiated from a 32-byte seed; the seed MUST have been generated with a cryptographically secure generator (this library does not include provisions for this generation step). Following RFC 8032, the seed is derived into a secret scalar, and an extra private value used for deterministic signature generation. The private key allows signature generation with the Ed25519, Ed25519ctx and Ed25519ph variants (in the third case, the pre-hashed message is provided by the caller). The PublicKey structure represents a public key, i.e. a curve point (and its 32-byte encoding as an additional field). Signature verification functions are available on PublicKey, again for Ed25519, Ed25519ctx and Ed25519ph.

Ed25519 Edge Cases

It is known that there is a great amount of variation about how existing Ed25519 implementations handle signatures and public keys which are slightly out of the strict standard formats, e.g. when non-canonical encodings are used. See Taming the many EdDSAs for some analysis. In the notations of that paper, this implementation is a strict RFC 8032 / FIPS 186-5 variant; it accepts test vectors 0 to 5, and reject vectors 6 to 11 (see table 5 in the paper). In other words:

Canonical encoding of both points and scalars is enforced. Non-canonical encodings are rejected.

The cofactored verification equation is used (i.e. including the multiplication by 8).

Points outside of the subgroup of order L, including low-order points and the identity point, are accepted both for the R component of the signatures, and for public keys.

The S component of a signature is accepted as long as it is canonically encoded (i.e. in the 0 to L-1 range). Zero is accepted. The full 32 bytes are used: the three top bits of the last byte, though always of value 0, are checked.

Truncated Signatures

The PublicKey::verify_trunc_*() functions support truncated signatures: a 64-byte signature is provided, but the last few bits are considered to have been reused for encoding other data, and thus are ignored. The verification function then tries to recompute the complete, original, untruncated signature. This process is safe since neither truncation nor reconstruction involve usage of the private key, and the original signature is obtained as an outcome of the process. Up to 32 bits (i.e. four whole bytes) can be rebuilt by this implementation, which corresponds to shrinking the minimum signature encoding size from 64 down to 60 bytes.

Signature reconstruction cost increases with the number of ignored bits; when 32 bits are ignored, the verification cost is about 25 to 35 times the cost of verifying an untruncated signature.

Structs

Point

A point on the twisted Edwards curve edwards25519.

PrivateKey

An Ed25519 private key.

PublicKey

An Ed25519 public key.

Type Aliases

Scalar

Integers modulo L = 2^252 + 27742317777372353535851937790883648493.