Struct crrl::frost::p256::SignerPrivateKeyShare

pub struct SignerPrivateKeyShare {
    pub ident: Scalar,
    /* private fields */
}

A private key share.

This structure contains the private key share of a given signer. It includes the signer’s identifier, private key (a scalar), corresponding public key (a point), and group public key. The signer’s public key is recomputed from the signer’s private key when decoding, so it always matches the signers private key.

Fields
ident: Scalar

Signer identifier

Implementations

impl SignerPrivateKeyShare

pub const ENC_LEN: usize = 97usize

Private key share encoded length (in bytes).

pub fn encode(self) -> [u8; 97]

Encodes this private key share into bytes.

pub fn decode(buf: &[u8]) -> Option<Self>

Decodes this share from bytes.

The process fails (i.e. returns None) if the source slice does not have a proper length or does not contain properly canonical encodings of the share identifier, share of the private key, or group public key.

pub fn get_public_key(self) -> SignerPublicKey

Get the public key for this signer.

pub fn verify_split(self, vsscomm: &[VSSElement]) -> bool

Verifies that this share was properly computed, given the VSS commitments.

This function is called vss_verify in the FROST specification.

pub fn commit<T: CryptoRng + RngCore>(self, rng: &mut T) -> (Nonce, Commitment)

Generates nonces and commitments for a new signature generation.

The returned Nonce and Commitment should be remembered by the signer for round 2. The Commitment should be sent to the coordinator (Nonce is secret and MUST NOT be revealed to anybody).

pub fn sign( self, nonce: Nonce, comm: Commitment, msg: &[u8], commitment_list: &[Commitment], ) -> Option<SignatureShare>

Computes a signature share.

The nonce and commitment (previously generated with a commit() call, the commitment was sent to the coordinator) is combined with the message and list of signer commitments selected by the coordinator.

This function may fail if the list of commitments is too short (less than two commitments), or not in the expected order (by ascending signer identifier), or contains duplicates (two commitments with the same identifier), or does not contain this signer’s identifier, or contains this signer’s identifier but with a different commitment. In all failure cases, None is returned.

The signer’s own nonce (nonce) and commitment (comm) MUST match each other.