Struct crrl::jq255e::Point

pub struct Point { /* private fields */ }

An element in the jq255e group.

Implementations

impl Point

pub const NEUTRAL: Self = _

The group neutral element.

pub const BASE: Self = _

The conventional base point (group generator).

This point generates the whole group, which has prime order r (integers modulo r are represented by the Scalar type).

pub fn set_decode(&mut self, buf: &[u8]) -> u32

Tries to decode a point from bytes.

If the source slice has not length exactly 32 bytes, then decoding fails. If the source bytes are not a valid, canonical encoding of a group element, then decoding fails. On success, 0xFFFFFFFF is returned; on failure, 0x00000000 is returned. On failure, this point is set to the neutral.

If the source length is exactly 32 bytes, then the decoding outcome (success or failure) should remain hidden from timing-based side channels.

pub fn decode(buf: &[u8]) -> Option<Point>

Tries to decode a point from some bytes.

Decoding succeeds only if the source slice has length exactly 32 bytes, and contains the canonical encoding of a valid curve point. Sicne this method returns an Option<Point>, it inherently leaks (through timing-based side channels) whether decoding succeeded or not; to avoid that, consider using set_decode(). The decoded point itself, however, does not leak.

pub fn encode(self) -> [u8; 32]

Encodes this point into exactly 32 bytes.

Encoding is always canonical.

pub fn set_double(&mut self)

Doubles this point (in place).

pub fn double(self) -> Self

Doubles this point.

pub fn set_xdouble(&mut self, n: u32)

Doubles this point n times (in place).

pub fn xdouble(self, n: u32) -> Self

Doubles this point n times.

pub fn set_neg(&mut self)

Negates this point (in place).

pub fn set_mul_small(&mut self, n: u64)

Multiplies this point by a small integer.

This operation is constant-time with regard to the source point, but NOT with regard to the multiplier; the multiplier n MUST NOT be secret.

pub fn equals(self, rhs: Self) -> u32

Compares two points for equality.

Returned value is 0xFFFFFFFF if the two points are equal, 0x00000000 otherwise.

pub fn isneutral(self) -> u32

Tests whether this point is the neutral (identity point on the curve).

Returned value is 0xFFFFFFFF for the neutral, 0x00000000 otherwise.

pub fn set_cond(&mut self, P: &Self, ctl: u32)

Conditionally copies the provided point (P) into self.

If ctl = 0xFFFFFFFF, then the value of P is copied into self.

If ctl = 0x00000000, then the value of self is unchanged.

ctl MUST be equal to 0x00000000 or 0xFFFFFFFF.

pub fn select(P0: &Self, P1: &Self, ctl: u32) -> Self

Returns a point equal to P0 (if ctl = 0x00000000) or P1 (if ctl = 0xFFFFFFFF).

Value ctl MUST be either 0x00000000 or 0xFFFFFFFF.

pub fn set_condneg(&mut self, ctl: u32)

Conditionally negates this point.

This point is negated if ctl = 0xFFFFFFFF, but kept unchanged if ctl = 0x00000000. ctl MUST be equal to 0x00000000 or 0xFFFFFFFF.

pub fn hash_to_curve(hash_name: &str, data: &[u8]) -> Self

Hashes some data into a point.

Given some input bytes, a group element is deterministically generated; the output distribution should be indistinguishable from uniform random generation, and the discrete logarithm of the output relatively to any given point is unknown.

The input bytes are provided as data. If these bytes are a hash value, then the hash function name should be provided as hash_name, corresponding to one of the defined constants (HASHNAME_SHA256, HASHNAME_BLAKE2S, etc). In general, the name to use is the “formal” name of the hash function, converted to lowercase and without punctuation signs (e.g. SHA-256 uses the name sha256). If the input bytes are not an already computed hash value, but some other raw data, then hash_name shall be set to an empty string.

pub const HASHNAME_SHA224: &'static str = "sha224"

pub const HASHNAME_SHA256: &'static str = "sha256"

pub const HASHNAME_SHA384: &'static str = "sha384"

pub const HASHNAME_SHA512: &'static str = "sha512"

pub const HASHNAME_SHA512_224: &'static str = "sha512224"

pub const HASHNAME_SHA512_256: &'static str = "sha512256"

pub const HASHNAME_SHA3_224: &'static str = "sha3224"

pub const HASHNAME_SHA3_256: &'static str = "sha3256"

pub const HASHNAME_SHA3_384: &'static str = "sha3384"

pub const HASHNAME_SHA3_512: &'static str = "sha3512"

pub const HASHNAME_BLAKE2B: &'static str = "blake2b"

pub const HASHNAME_BLAKE2S: &'static str = "blake2s"

pub const HASHNAME_BLAKE3: &'static str = "blake3"

pub fn set_mul(&mut self, n: &Scalar)

Multiplies this point by a scalar (in place).

This operation is constant-time with regard to both the points and the scalar value.

pub fn set_mulgen(&mut self, n: &Scalar)

Sets this point by multiplying the conventional generator by the provided scalar.

This operation is constant-time. It is faster than using the generic multiplication on Self::BASE.

pub fn mulgen(n: &Scalar) -> Self

Creates a point by multiplying the conventional generator by the provided scalar.

This operation is constant-time.

pub fn set_mul_add_mulgen_vartime(&mut self, u: &Scalar, v: &Scalar)

Given scalars u and v, sets this point to u*self + v*B (with B being the conventional generator of the prime order subgroup).

This can be used to support Schnorr signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn mul_add_mulgen_vartime(self, u: &Scalar, v: &Scalar) -> Self

Given scalars u and v, returns u*self + v*B (with B being the conventional generator of the prime order subgroup).

This can be used to support Schnorr signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn set_mul128_add_mulgen_vartime(&mut self, u: u128, v: &Scalar)

Given integer u and scalar v, sets this point to u*self + v*B (with B being the conventional generator of the prime order subgroup).

This can be used to support Schnorr signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn mul128_add_mulgen_vartime(self, u: u128, v: &Scalar) -> Self

Given integer u and scalar v, returns u*self + v*B (with B being the conventional generator of the prime order subgroup).

This can be used to support Schnorr signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.