Module crrl::ristretto255

Ristreto255 implementation.

The ristretto255 group is a prime order group currently specified in draft-irtf-cfrg-ristretto255-decaf448-07. It is internally defined over the curve Edwards25519, which is a twisted Edwards curve. Users of ristretto255 should not, in general, think about the underlying curve points; the group has prime order and that is the abstraction that is convenient for building cryptographic protocols.

The Point structure represents a ristretto255 point. Such points can be encoded into 32 bytes, and decoded back; encoding is always canonical, and this is enforced upon decoding. This implementation strictly follows the draft; in particular, when decoding a point from 32 bytes, it verifies that the top bit (most significant bit of the last byte) is zero (this bit is always zero for a valid encoding, but the code does not ignore it when decoding).

The Scalar type is an alias for the ed25519::Scalar type, which represents integers modulo the ristretto255 order L (in the twisted Edwards curve, this is the order of a specific subgroup of the curve).

Structs

Point

A ristretto255 point.

Type Aliases

Scalar

A ristretto255 scalar (integer modulo the group prime order L).