Struct crrl::ristretto255::Point

pub struct Point(/* private fields */);

A ristretto255 point.

Implementations

impl Point

pub const NEUTRAL: Self = _

The neutral element (identity point) in the group.

pub const BASE: Self = _

The conventional base point in the group.

pub fn set_decode(&mut self, buf: &[u8]) -> u32

Sets this element by decoding its binary representation.

If the input does not have length exactly 32 bytes, or if the input has length 32 bytes but is not the valid, canonical encoding of a ristretto255 point, then this function sets self to the neutral element and returns 0x00000000; otherwise, it sets self to the decoded element and returns 0xFFFFFFFF.

pub fn decode(buf: &[u8]) -> Option<Self>

Decodes an element from its binary representation.

If the input does not have length exactly 32 bytes, or if the input has length 32 bytes but is not the valid, canonical encoding of a ristretto255 point, then this function returns None. Otherwise, it returns the decoded element.

Since this function uses an option type, outsiders may detect through side-channels whether decoding succeeded or failed; however, the decoded value should not leak.

pub fn encode(self) -> [u8; 32]

Encodes this element into bytes.

Encoding is always canonical.

pub fn equals(self, rhs: Self) -> u32

Compares two points for equality.

Returned value is 0xFFFFFFFF if the two points are equal, 0x00000000 otherwise.

Note: this function is vastly faster than encoding the two elements and comparing the two encodings.

pub fn isneutral(self) -> u32

Tests whether this element is the neutral (identity point).

Returned value is 0xFFFFFFFF for the neutral, 0x00000000 for all other elements.

pub fn set_cond(&mut self, P: &Self, ctl: u32)

Conditionally copies the provided element (P) into self.

If ctl = 0xFFFFFFFF, then the value of P is copied into self.

If ctl = 0x00000000, then the value of self is unchanged.

Value ctl MUST be equal to either 0x00000000 or 0xFFFFFFFF.

pub fn select(P0: &Self, P1: &Self, ctl: u32) -> Self

Returns an element equal to P0 (if ctl = 0x00000000) or to P1 (if ctl = 0xFFFFFFFF).

Value ctl MUST be equal to either 0x00000000 or 0xFFFFFFFF.

pub fn set_condneg(&mut self, ctl: u32)

Conditionally negates this point.

This point is negated if ctl = 0xFFFFFFFF, but kept unchanged if ctl = 0x00000000.

Value ctl MUST be equal to either 0x00000000 or 0xFFFFFFFF.

pub fn one_way_map(buf: &[u8]) -> Self

The one-way map of bytes to ristretto255 elements.

This is the map described in the draft, section 4.3.4. The input MUST have length exactly 64 bytes (a panic is triggered otherwise). If the input is itself a 64-byte output of a secure hash function (e.g. SHA-512) then this constitutes a hash function with output in ristretto255 (output is then indistinguishable from random uniform selection).

pub fn set_neg(&mut self)

Negates this element.

pub fn set_mulgen(&mut self, n: &Scalar)

Sets this element to n times the conventional base (Self::BASE).

pub fn mulgen(n: &Scalar) -> Self

Returns the product of the conventional base (Self::BASE) by the provided scalar.

pub fn set_double(&mut self)

Doubles this element (in place).

pub fn double(self) -> Self

Doubles this element.

pub fn set_xdouble(&mut self, n: u32)

Doubles this element n times (in place).

pub fn xdouble(self, n: u32) -> Self

Doubles this element n times.

pub fn mul_add_mulgen_vartime(self, u: &Scalar, v: &Scalar) -> Self

Given scalars u and v, returns u*self + v*B (with B being the conventional generator of the prime order subgroup).

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn verify_helper_vartime(self, R: &Point, s: &Scalar, k: &Scalar) -> bool

Check whether s*B = R + k*A, for the provided scalars s and k, provided points A (self) and R, and conventional generator B.

Returned value is true on match, false otherwise. This function is meant to support EdDSA-style signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.