Struct crrl::secp256k1::Point

pub struct Point { /* private fields */ }

A point on the short Weierstraß curve secp256k1.

Implementations

impl Point

pub const NEUTRAL: Self = _

The neutral element (point-at-infinity) in the curve.

pub const BASE: Self = _

The conventional base point in the curve.

Like all non-neutral points in secp256k1, it generates the whole curve.

pub fn set_decode(&mut self, buf: &[u8]) -> u32

Tries to decode a point.

This function accepts the following encodings and lengths:

A single byte of value 0x00: the point-at-infinity.

A byte of value 0x02 or 0x03, followed by exactly 32 bytes (unsigned big-endian encoding of the x coordinate): compressed encoding of a non-neutral point.

A byte of value 0x04, followed by exactly 64 bytes (unsigned big-endian encodings of x and y): uncompressed encoding of a non-neutral point.

The (very rarely encountered) “hybrid” encoding (like uncompressed, but the least significant bit of y is also copied into the first byte, which has value 0x06 or 0x07) is not supported.

On success, this structure is set to the decoded point, and 0xFFFFFFFF is returned. On failure, this structure is set to the neutral point, and 0x00000000 is returned. A failure is reported if the coordinates can be decoded but do not correspond to a point on the curve.

Constant-time behaviour: timing-based side channels may leak which encoding type was used (neutral, compressed, uncompressed) but not the value of the obtained point, nor whether the encoding was for a valid point.

pub fn decode(buf: &[u8]) -> Option<Point>

Tries to decode a point.

This function accepts the following encodings and lengths:

A single byte of value 0x00: the point-at-infinity.

A byte of value 0x02 or 0x03, followed by exactly 32 bytes (unsigned big-endian encoding of the x coordinate): compressed encoding of a non-neutral point.

A byte of value 0x04, followed by exactly 64 bytes (unsigned big-endian encodings of x and y): uncompressed encoding of a non-neutral point.

The (very rarely encountered) “hybrid” encoding (like uncompressed, but the least significant bit of y is also copied into the first byte, which has value 0x06 or 0x07) is not supported.

On success, the decoded point is returned; on failure, None is returned. A failure is reported if the coordinates can be decoded but do not correspond to a point on the curve.

Constant-time behaviour: timing-based side channels may leak which encoding type was used (neutral, compressed, uncompressed) but not the value of the obtained point, nor whether the encoding was for a valid point.

pub fn encode_compressed(self) -> [u8; 33]

Encodes this point in compressed format (33 bytes).

If the point is the neutral then [0u8; 33] is returned, which is NOT the standard encoding of the neutral (standard is a single byte of of value 0x00); for a non-neutral point, the first byte is always equal to 0x02 or 0x03, never to 0x00.

pub fn encode_uncompressed(self) -> [u8; 65]

Encodes this point in uncompressed format (65 bytes).

If the point is the neutral then [0u8; 65] is returned, which is NOT the standard encoding of the neutral (standard is a single byte of of value 0x00); for a non-neutral point, the first byte is always equal to 0x04, never to 0x00.

pub fn to_affine(self) -> (GFsecp256k1, GFsecp256k1, u32)

Gets the affine (x, y) coordinates for this point.

Values (x, y, r) are returned, with x and y being field elements, and r a u32 value that qualifies the outcome:

if the point is the neutral, then x = 0, y = 0 and r = 0x00000000;

otherwise, x and y are the affine coordinates, and r = 0xFFFFFFFF.

Note that there is no point with x = 0 or with y = 0 on the curve.

pub fn to_projective(self) -> (GFsecp256k1, GFsecp256k1, GFsecp256k1)

Gets the projective coordinates (X:Y:Z) for this point.

Values (X, Y, Z) are returned, such that:

if the point is the neutral (point-at-infinity), then X and Z are 0;

otherwise, Z != 0, and the affine point coordinates are x = X/Z and y = Y/Z.

By definition, projective coordinates for a given point are not unique; two equal points may have distinct projective coordinates.

The Y coordinate is never 0. The X coordinate may be 0 only for the neutral point.

pub fn set_affine(&mut self, x: GFsecp256k1, y: GFsecp256k1) -> u32

Sets this instance from the provided affine coordinates.

If the coordinates designate a valid curve point, then the function returns 0xFFFFFFFF; otherwise, this instance is set to the neutral, and the function returns 0x00000000.

pub fn from_affine(x: GFsecp256k1, y: GFsecp256k1) -> Option<Self>

Creates an instance from the provided affine coordinates.

The coordinates are verified to comply with the curve equation; if they do not, then None is returned.

Note: whether the point is on the curve or not may leak through side channels; however, the actual value of the point should not leak.

pub fn set_projective( &mut self, X: GFsecp256k1, Y: GFsecp256k1, Z: GFsecp256k1, ) -> u32

Sets this instance from the provided projective coordinates.

If the coordinates designate a valid curve point, then the function returns 0xFFFFFFFF; otherwise, this instance is set to the neutral, and the function returns 0x00000000.

This function accepts any (X:Y:0) triplet as a representation of the point-at-infinity.

pub fn from_projective( X: GFsecp256k1, Y: GFsecp256k1, Z: GFsecp256k1, ) -> Option<Self>

Creates an instance from the provided projective coordinates.

The coordinates are verified to comply with the curve equation; if they do not, then None is returned.

This function accepts any (X:Y:0) triplet as a representation of the point-at-infinity.

Note: whether the point is on the curve or not may leak through side channels; however, the actual value of the point should not leak.

pub fn set_double(&mut self)

Doubles this point (in place).

This function is somewhat faster than using plain point addition.

pub fn double(self) -> Self

Doubles this point.

This function is somewhat faster than using plain point addition.

pub fn set_xdouble(&mut self, n: u32)

Doubles this point n times (in place).

pub fn xdouble(self, n: u32) -> Self

Doubles this point n times.

When n > 1, this function is faster than calling double() n times.

pub fn set_neg(&mut self)

Negates this point (in place).

pub fn set_mul_small(&mut self, n: u64)

Multiplies this point by a small integer.

This operation is constant-time with regard to the source point, but NOT with regard to the multiplier; the multiplier n MUST NOT be secret.

pub fn equals(self, rhs: Self) -> u32

Compares two points for equality.

Returned value is 0xFFFFFFFF if the two points are equal, 0x00000000 otherwise.

pub fn isneutral(self) -> u32

Tests whether this point is the neutral (point-at-infinity).

Returned value is 0xFFFFFFFF for the neutral, 0x00000000 otherwise.

pub fn set_cond(&mut self, P: &Self, ctl: u32)

pub fn select(P0: &Self, P1: &Self, ctl: u32) -> Self

Returns a point equal to P0 (if ctl = 0x00000000) or P1 (if ctl = 0xFFFFFFFF).

Value ctl MUST be either 0x00000000 or 0xFFFFFFFF.

pub fn set_condneg(&mut self, ctl: u32)

Conditionally negates this point.

This point is negated if ctl = 0xFFFFFFFF, but kept unchanged if ctl = 0x00000000.

Value ctl MUST be either 0x00000000 or 0xFFFFFFFF.

pub fn set_mul(&mut self, n: &Scalar)

Multiplies this point by a scalar (in place).

This operation is constant-time with regard to both the points and the scalar value.

pub fn set_mulgen(&mut self, n: &Scalar)

Sets this point by multiplying the conventional generator by the provided scalar.

This operation is constant-time. It is faster than using the generic multiplication on Self::BASE.

pub fn mulgen(n: &Scalar) -> Self

Creates a point by multiplying the conventional generator by the provided scalar.

This operation is constant-time. It is faster than using the generic multiplication on Self::BASE.

pub fn set_mul_add_mulgen_vartime(&mut self, u: &Scalar, v: &Scalar)

Given scalars u and v, sets this point to u*self + v*G (with G being the conventional generator point, aka Self::BASE).

This function can be used to support ECDSA signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn mul_add_mulgen_vartime(self, u: &Scalar, v: &Scalar) -> Self

Given scalars u and v, returns point u*self + v*G (with G being the conventional generator point, aka Self::BASE).

This function can be used to support ECDSA signature verification.

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.

pub fn verify_helper_vartime(self, R: &Point, s: &Scalar, k: &Scalar) -> bool

Check whether s*G = R + k*Q, for the provided scalars s and k, provided points Q (self) and R, and conventional generator G.

Returned value is true on match, false otherwise. This function is meant to support Schnorr signature verification (e.g. as defined in FROST).

THIS FUNCTION IS NOT CONSTANT-TIME; it shall be used only with public data.