Function crrl::x25519::x25519

pub fn x25519(point: &[u8; 32], scalar: &[u8; 32]) -> [u8; 32]

X25519 function (from RFC 7748), general case.

The source point is provided as an array of 32 bytes (point), as well as the scalar (scalar). In RFC 7748 terminology, the point parameter is the little-endian encoding of the u coordinate of a point on the Montgomery curve or on its quadratic twist, and the scalar parameter is the little-endian encoding of the scalar. The function “clamps” the scalar (bits 0, 1, 2 and 255 are cleared, bit 254 is set) then interprets the clamped scalar as an integer (little-endian convention), with which the provided curve point is multiplied; the u coordinate of the resulting point is then encoded and returned.